Refactor PropertyTreeQueue class
